requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
numpy>=1.24
pillow>=10.0
opencv-python-headless>=4.8
keras>=3.2
tensorflow>=2.15
python-multipart>=0.0.9